:recycle: refactor(#20): enumeration 일괄 core에서 관리하기 위해 리팩터링
